Change everything to use new AnalyzedString
The id is now a magic string type that is filled by running
the `browse_key_analyzer` against the term/locID passed in. We
no longer need to do any sort of client-side analysis for
passing to solr, although we're still using the ICU
library during attempts to match the actual extracted
terms with the Library of Congress name authority file.