feat: can generate remediated authorities file
Adds method for generating the conf/remediated_subjects.xml file. It
pulls the records from Alma and writes them to that conf file. In the
browse cli, the method `subjects generate_remediated_authorities_file`
is added that configuration pulling and updating.

A github action is added to call the `browse subjects
generate_remediate_authorities_file` cli command and generate a PR if
the content of the file has changed.