fix: changes max_queue to 0
The term fetchers were not retrieving all of the terms from solr because
the thread queue was overloaded. The number of threads is bounded for
these processes and it's important that all of the terms are collected,
so this value is set to 0 which means infinity.

Also changes ThreadPoolExecutor to FixedThreadPool in
authority_browse term_fetcher so that they match.